+++ app/javascript/entrypoints/inertia.ts
@@ -1,25 +1,13 @@
 import { createInertiaApp } from '@inertiajs/react'
-import { createElement, ReactNode } from 'react'
+import { createElement, type ReactNode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './application.css'
 
-// Temporary type definition, until @inertiajs/react provides one
 type ResolvedComponent = {
   default: ReactNode
   layout?: (page: ReactNode) => ReactNode
 }
 
 createInertiaApp({
-  // Set default page title
-  // see https://inertia-rails.dev/guide/title-and-meta
-  //
-  // title: title => title ? `${title} - App` : 'App',
-
-  // Disable progress bar
-  //
-  // see https://inertia-rails.dev/guide/progress-indicators
-  // progress: false,
-
   resolve: (name) => {
     const pages = import.meta.glob<ResolvedComponent>('../pages/**/*.tsx', {
       eager: true,
@@ -28,13 +16,6 @@ createInertiaApp({
     if (!page) {
       console.error(`Missing Inertia page component: '${name}.tsx'`)
     }
-
-    // To use a default layout, import the Layout component
-    // and use the following line.
-    // see https://inertia-rails.dev/guide/pages#default-layouts
-    //
-    // page.default.layout ||= (page) => createElement(Layout, null, page)
-
     return page
   },
 
@@ -43,9 +24,7 @@ createInertiaApp({
       createRoot(el).render(createElement(App, props))
     } else {
       console.error(
-        'Missing root element.\n\n' +
-          'If you see this error, it probably means you load Inertia.js on non-Inertia pages.\n' +
-          'Consider moving <%= vite_typescript_tag "inertia" %> to the Inertia-specific layout instead.',
+        'Missing root element. Move `vite_typescript_tag "inertia"` into an Inertia-specific layout.',
       )
     }
   },