+++ db/migrate/20260423000004_create_users.rb
@@ -4,6 +4,7 @@ class CreateUsers < ActiveRecord::Migration[8.0]
       t.string :email, null: false
       t.string :password_digest, null: false
       t.string :timezone
+      t.boolean :admin, null: false, default: false
 
       t.timestamps
     end