+++ db/seeds.rb
@@ -7,3 +7,15 @@
 #   ["Action", "Comedy", "Drama", "Horror"].each do |genre_name|
 #     MovieGenre.find_or_create_by!(name: genre_name)
 #   end
+
+User.find_or_create_by!(email: "user@test.com") do |user|
+  user.password = "test123"
+  user.timezone = "America/New_York"
+  user.admin = false
+end
+
+User.find_or_create_by!(email: "admin@test.com") do |user|
+  user.password = "test123"
+  user.timezone = "America/New_York"
+  user.admin = true
+end